Module epp_proxy::client

Async/await EPP client.

Messages should be injected into the server using the helper functions in subordinate modules such as contact, host, and domain.

Re-exports

pub use router::CommandResponse;

pub use router::RequestMessage;

pub use router::RequestSender;

pub use router::Response;

pub use router::Sender;

Modules

balance
Commands relating to balance enquiries

contact
Commands relating to contact objects

domain
EPP commands relating to domain objects

epp

epp_like

eurid
EPP commands relating to EURid extensions

fee

host
EPP commands relating to host (nameserver) objects

launch

maintenance
EPP commands relating to draft-ietf-regext-epp-registry-maintenance

nominet
EPP commands relating to nominet specific features

poll
EPP commands relating to domain objects

rgp
EPP commands relating to nominet specific features

router

tmch

traficom

verisign

Structs

ClientConf

LogoutRequest

ServerFeatures
Features supported by the server

Enums

ClientCertConf

Error
Possible errors returned by the EPP client

TransferStatus

Functions

logout
Ends an EPP session

send_epp_client_request