Module epp_proxy::client::domain

EPP commands relating to domain objects

Structs

CheckRequest

CheckResponse
Response to a domain check query

ClaimsCheckRequest

ClaimsCheckResponse
Response to a domain claims check query

CreateData

CreateInfo

CreateRequest

CreateResponse

DeleteRequest

DeleteResponse

InfoContact
Additional contact associated with a domain

InfoRequest

InfoResponse
Response to a domain info query

PanData

Period
Domain registration period

RenewData

RenewRequest

RenewResponse

SecDNSDSData

SecDNSData
DNSSEC key data

SecDNSKeyData

TrademarkCheckRequest

TransferAcceptRejectRequest

TransferData

TransferQueryRequest

TransferRequestRequest

TransferResponse

UpdateInfo

UpdateRequest

UpdateResponse

UpdateSecDNS

VerisignSyncRequest

Enums

InfoHost

InfoNameserver
Nameserver associated with a domain

PeriodUnit
Domain registration period time unit

SecDNSDataType

Status

UpdateObject

UpdateSecDNSRemove

Functions

check
Checks if a domain name is available

create
Registers a new domain

delete
Deletes a domain name

info
Fetches information about a domain name

launch_claims_check
Checks if a domain name has claims registered for a launch phase

launch_trademark_check
Checks if a domain name has trademarks registered for it

renew
Renews a domain name

transfer_accept
Accepts the transfer of a domain name

transfer_cancel
Cancels the pending transfer of a domain name

transfer_query
Queries the current transfer status of a domain name

transfer_reject
Rejects the transfer of a domain name

transfer_request
Requests the transfer of a domain name

update
Updates properties of a domain name

verisign_sync
Performs a Verisign ConsoliDate