Struct epp_proxy::client::domain::InfoResponse

pub struct InfoResponse {

    pub name: String,
    pub registry_id: String,
    pub statuses: Vec<Status>,
    pub registrant: String,
    pub contacts: Vec<InfoContact>,
    pub nameservers: Vec<InfoNameserver>,
    pub hosts: Vec<String>,
    pub client_id: String,
    pub client_created_id: Option<String>,
    pub creation_date: Option<DateTime<Utc>>,
    pub expiry_date: Option<DateTime<Utc>>,
    pub last_updated_client: Option<String>,
    pub last_updated_date: Option<DateTime<Utc>>,
    pub last_transfer_date: Option<DateTime<Utc>>,
    pub rgp_state: Vec<RGPState>,
    pub auth_info: Option<String>,
    pub sec_dns: Option<SecDNSData>,
    pub launch_info: Option<LaunchInfoData>,
    pub donuts_fee_data: Option<DonutsFeeData>,
    pub whois_info: Option<InfoWhois>,
    pub isnic_info: Option<DomainInfo>,
    pub eurid_data: Option<DomainInfo>,
    pub eurid_idn: Option<Idn>,

}

Response to a domain info query

Fields
name: String

Domain name in question

registry_id: String

Internal registry ID

statuses: Vec<Status>

Statuses attached to the domain

registrant: String

Contact ID of the registrant

contacts: Vec<InfoContact>

Additional contacts on the domain

nameservers: Vec<InfoNameserver>

Nameservers for the domain

hosts: Vec<String>

Host names attached to the domain

client_id: String

Sponsoring client ID

client_created_id: Option<String>

ID of the client that originally registered the domain

creation_date: Option<DateTime<Utc>>

Date of initial registration

expiry_date: Option<DateTime<Utc>>

Date of registration expiration

last_updated_client: Option<String>

ID of the client that last updated the domain

last_updated_date: Option<DateTime<Utc>>

Date of last update

last_transfer_date: Option<DateTime<Utc>>

Date of last transfer

rgp_state: Vec<RGPState>

Redemption grace period state of the domain

auth_info: Option<String>sec_dns: Option<SecDNSData>

DNSSEC data

launch_info: Option<LaunchInfoData>donuts_fee_data: Option<DonutsFeeData>whois_info: Option<InfoWhois>isnic_info: Option<DomainInfo>eurid_data: Option<DomainInfo>eurid_idn: Option<Idn>