Function epp_proxy::client::recv_msg

async fn recv_msg<'_, '_, '_, R: Unpin + AsyncRead>(
    sock: &'_ mut R, 
    host: &'_ str, 
    root: &'_ Path
) -> Result<EPPMessage, bool>

Attempts to read and decode an EPP message.

Reads from a tokio async reader in conformance with RFC 5734 for the binary message data, and decodes in conformance with RFC 5730 and related documents for said message. Will return Ok(proto::EPPMessage) on success or Err(()) on any error. In such cases the client should close the connection and restart.

Arguments

sock - A tokio async reader

host - Host name for error reporting

root - Root dir for storing message logs