Struct epp_proxy::client::EPPClient

pub struct EPPClient {
    host: String,
    tag: String,
    password: String,
    server_id: String,
    ready: bool,
    router: Router,
    features: EPPClientServerFeatures,
}

Main client struct for the EEP client

Fields
host: Stringtag: Stringpassword: Stringserver_id: Stringready: bool

Is the EPP server in a state to receive and process commands

router: Routerfeatures: EPPClientServerFeatures

What features does the server support

Methods

impl EPPClient

pub fn new(host: &str, tag: &str, password: &str) -> Self

Creates a new EPP client ready to be started

Arguments

host - The server connection string, in the form domain:port

tag - The client ID/tag to login with

password - The password to login with

pub fn start(self) -> Sender<Request>

Starts up the EPP server and returns the sending end of a tokio channel to inject commands into the client to be processed

async fn _main_loop<'_>(&'_ mut self, receiver: Receiver<Request>)

async fn _handle_response<'_>(&'_ mut self, res: EPPMessage) -> Result<(), ()>

async fn _setup_connection<'_, '_>(
&'_ mut self,
sock: &'_ mut TlsStream<TcpStream>
) -> Result<(), bool>

async fn _process_greeting<'_>(
&'_ mut self,
greeting: EPPGreeting
) -> Result<(), ()>

async fn _login<'_, '_>(
&'_ self,
sock: &'_ mut TlsStream<TcpStream>
) -> Result<(), ()>

async fn _send_command<'_, '_, W: Unpin + AsyncWrite, M: Into<Option<Uuid>>>(
&'_ self,
command: EPPCommandType,
sock: &'_ mut W,
message_id: M
) -> Result<Uuid, ()>

async fn _send_msg<'_, '_, '_, W: Unpin + AsyncWrite>(
&'_ self,
message: &'_ EPPMessage,
sock: &'_ mut W
) -> Result<(), ()>

async fn _close<'_, '_>(&'_ mut self, sock: &'_ mut TlsStream<TcpStream>)

async fn _connect<'_>(&'_ self) -> TlsStream<TcpStream>

async fn _try_connect<'_>(&'_ self) -> Result<TlsStream<TcpStream>, ()>