Struct epp_proxy::grpc::epp_proto::domain::DomainInfoReply

pub struct DomainInfoReply {

    pub name: String,
    pub registry_id: String,
    pub statuses: Vec<i32>,
    pub registrant: String,
    pub contacts: Vec<Contact>,
    pub nameservers: Vec<NameServer>,
    pub hosts: Vec<String>,
    pub client_id: String,
    pub client_created_id: Option<String>,
    pub creation_date: Option<Timestamp>,
    pub expiry_date: Option<Timestamp>,
    pub last_updated_client: Option<String>,
    pub last_updated_date: Option<Timestamp>,
    pub last_transfer_date: Option<Timestamp>,
    pub registry_name: String,
    pub rgp_state: Vec<i32>,
    pub auth_info: Option<String>,
    pub sec_dns: Option<SecDnsData>,
    pub launch_info: Option<LaunchInfoData>,
    pub donuts_fee_data: Option<DonutsFeeData>,
    pub verisign_whois_info: Option<VerisignWhoisInfo>,
    pub cmd_resp: Option<CommandResponse>,
    pub eurid_idn: Option<Idn>,
    pub eurid_data: Option<DomainInfo>,
    pub isnic_info: Option<DomainInfo>,

}

Fields
name: Stringregistry_id: Stringstatuses: Vec<i32>registrant: Stringcontacts: Vec<Contact>nameservers: Vec<NameServer>hosts: Vec<String>client_id: Stringclient_created_id: Option<String>creation_date: Option<Timestamp>expiry_date: Option<Timestamp>last_updated_client: Option<String>last_updated_date: Option<Timestamp>last_transfer_date: Option<Timestamp>registry_name: Stringrgp_state: Vec<i32>auth_info: Option<String>sec_dns: Option<SecDnsData>launch_info: Option<LaunchInfoData>donuts_fee_data: Option<DonutsFeeData>verisign_whois_info: Option<VerisignWhoisInfo>cmd_resp: Option<CommandResponse>eurid_idn: Option<Idn>eurid_data: Option<DomainInfo>isnic_info: Option<DomainInfo>
Implementations

impl DomainInfoReply

pub fn statuses(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<DomainStatus>>

Returns an iterator which yields the valid enum values contained in statuses.

pub fn push_statuses(&mut self, value: DomainStatus)

Appends the provided enum value to statuses.

pub fn rgp_state(
&self
) -> FilterMap<Cloned<Iter<'_, i32>>, fn(_: i32) -> Option<RgpState>>

Returns an iterator which yields the valid enum values contained in rgp_state.

pub fn push_rgp_state(&mut self, value: RgpState)

Appends the provided enum value to rgp_state.