Module epp_proxy::xml_ser

Custom serde XML serializer

The special serde tag name $value equates to the inner value of an XML element. Tags starting with $attr: will be encoded as attributes rather than new elements. Namespaces and prefixes can be set using the tag name format {namespace}prefix:tag-name.

Structs

MapSerializer

SeqSerializer

Serializer

StructSerializer

StructVariantSerializer

Enums

_SerializerData

Functions

format_tag

to_string
Serialise serde item to XML