Crate epp_proxy

A proxy server for interacting with EPP servers over gRPC

The server will listen for gRPC requests on [::1]:50051 by deafault. See the proto/epp.proto file for information on the gRPC protobufs used to communicate with the server. Use --help to view more options.

Server expects configuration in json files it the folder ./conf/ relative to the programs current working directory on startup. JSON file should follow the structure of the ConfigFile struct, where id is a unique ID for identifying the register in gRPC commands, server is the TLS server to connect to in the form domain:port, tag is the client login ID, password is the client login password, new_password is the optional new EPP password if it is to be changed on login, zones is a list of DNS zones said server is responsible for such as ch, co.uk, and org.uk, client_cert is an optional TLS certificated bundle in PKCS12 format, pipelining defines support for multiple in flight commands, errata defines server errata.

Supported errata are:

traficom

verisign-tv

verisign-cc

rrpproxy

Example config file:

{
 "id": "nominet",
 "server": "ote-epp.nominet.org.uk:700",
 "tag": "AS207960",
 "new_password": "supersecretpassword",
 "password": "oldpassword",
 "zones": [
   "uk"
 ],
 "client_cert": "priv/as207960-registrar.pfx",
 "root_certs": ["root/uniregistry.pem"],
 "pipelining": true,
 "errata": "traficom"
}

Modules

built_info

client

Async/await EPP client.

grpc

Implements the gRPC interface for the EPP client

proto

Serde structs for serialisation and deserialisation of EPP XML messages (these are insane, stay away if you value your sanity)

Structs

AuthService

ConfigFile

HSMConfigFile

P11Engine

Holding type with drop for OpenSSL engine references

P11EngineInner

Router

Route requests to the correct EPP client for the authoritative registry

Enums

ClientCertConfig

ConfigServerType

Functions

cvt

cvt_p

main

oauth_client

server_identity