Module epp_proxy::client

Async/await EPP client.

Messages should be injected into the server using the helper functions in subordinate modules such as contact, host, and domain.

Re-exports

pub use router::CommandResponse;

pub use router::RequestMessage;

pub use router::RequestSender;

pub use router::Response;

pub use router::Sender;

Modules

balance

Commands relating to balance enquiries

contact

Commands relating to contact objects

dac

domain

EPP commands relating to domain objects

email_forward

EPP commands relating to email forwarding objects

epp

epp_like

eurid

EPP commands relating to EURid extensions

fee

host

EPP commands relating to host (nameserver) objects

isnic

launch

maintenance

EPP commands relating to draft-ietf-regext-epp-registry-maintenance

mark

nominet

EPP commands relating to nominet specific features

nominet_dac

personal_registration

poll

EPP commands relating to domain objects

rgp

EPP commands relating to nominet specific features

router

tmch

tmch_client

traficom

verisign

Structs

BlankRequest

ClientConf

NominetDACConf

Period

Phone

Enums

ClientCertConf

Error

Possible errors returned by the EPP client

PeriodUnit

TransferStatus

Traits

Client

Functions

logout

Ends an EPP session