Module epp_proxy::client

Async/await EPP client.

Messages should be injected into the server using the helper functions in subordinate modules such as [contact], [host], and [domain].

Re-exports

pub use router::Request;

pub use router::Response;

Modules

contact
EPP commands relating to contact objects

domain
EPP commands relating to domain objects

host
EPP commands relating to host (nameserver) objects

router
Routes requests into and out of the EPP client by keeping track of

Structs

EPPClient
Main client struct for the EEP client

EPPClientReceiver
Tokio task that attemps to read in EPP messages and push them onto a tokio channel as received.

EPPClientServerFeatures
Features supported by the EPP server

Enums

Error
Possible errors returned by the EPP client

Functions

recv_msg
Attempts to read and decode an EPP message.

send_epp_client_request

Type Definitions

Sender